moved write_h5 to hdf5imagingextractor
